+++ src/main/java/com/freshbourne/io/PagePointer.java
@@ -24,7 +24,7 @@ package com.freshbourne.io;
 public class PagePointer {
 	
 	private int offset;
-	private long id;
+	private int id;
 	
 	
 	
@@ -32,7 +32,7 @@ public class PagePointer {
 	 * @param offset
 	 * @param id
 	 */
-	public PagePointer(long id, int offset) {
+	public PagePointer(int id, int offset) {
 		super();
 		this.offset = offset;
 		this.id = id;
@@ -59,7 +59,7 @@ public class PagePointer {
 	/**
 	 * @return the pageId
 	 */
-	public long getId() {
+	public int getId() {
 		return id;
 	}
